+++ addons/gamification/models/badge.py
@@ -4,9 +4,10 @@
 import logging
 from datetime import date
 
-from odoo import api, fields, models, _, exceptions
+from odoo import api, tools, fields, models, _, exceptions
 
 _logger = logging.getLogger(__name__)
+
 
 class BadgeUser(models.Model):
     """User having received a badge"""
@@ -22,6 +23,9 @@ class BadgeUser(models.Model):
     challenge_id = fields.Many2one('gamification.challenge', string='Challenge originating', help="If this badge was rewarded through a challenge")
     comment = fields.Text('Comment')
     badge_name = fields.Char(related='badge_id.name', string="Badge Name", readonly=False)
+    level = fields.Selection(
+        [('bronze', 'Bronze'), ('silver', 'Silver'), ('gold', 'Gold')],
+        string='Badge Level', related="badge_id.level", store=True, readonly=True)
 
     def _send_badge(self):
         """Send a notification to a user for receiving a badge
@@ -66,7 +70,20 @@ class GamificationBadge(models.Model):
     name = fields.Char('Badge', required=True, translate=True)
     active = fields.Boolean('Active', default=True)
     description = fields.Text('Description', translate=True)
-    image = fields.Binary("Image", help="This field holds the image used for the badge, limited to 256x256")
+    image = fields.Binary("Image", help="This field holds the image used for the badge.")
+    image_medium = fields.Binary(
+        "Medium-sized badge image",
+        help="Medium-sized image of the badge. It is automatically "
+             "resized as a 128x128px image, with aspect ratio preserved. "
+             "Use this field in form views or some kanban views.")
+    image_small = fields.Binary(
+        "Small-sized badge image",
+        help="Small-sized image of the badge. It is automatically "
+             "resized as a 64x64px image, with aspect ratio preserved. "
+             "Use this field anywhere a small image is required.")
+    level = fields.Selection([
+        ('bronze', 'Bronze'), ('silver', 'Silver'), ('gold', 'Gold')],
+        string='Forum Badge Level', default='bronze')
 
     rule_auth = fields.Selection([
             ('everyone', 'Everyone'),
@@ -119,6 +136,17 @@ class GamificationBadge(models.Model):
     remaining_sending = fields.Integer(
         "Remaining Sending Allowed", compute='_remaining_sending_calc',
         help="If a maximum is set")
+
+    @api.model_create_multi
+    def create(self, values_list):
+        for vals in values_list:
+            tools.image_resize_images(vals)
+        return super(GamificationBadge, self).create(values_list)
+
+    @api.multi
+    def write(self, vals):
+        tools.image_resize_images(vals)
+        return super(GamificationBadge, self).write(vals)
 
     @api.depends('owner_ids')
     def _get_owners_info(self):
