+++ addons/gamification/models/res_users.py
@@ -1,0 +1,44 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import api, fields, models
+
+
+class Users(models.Model):
+    _inherit = 'res.users'
+
+    karma = fields.Integer('Karma', default=0)
+    badge_ids = fields.One2many('gamification.badge.user', 'user_id', string='Badges', copy=False)
+    gold_badge = fields.Integer('Gold badges count', compute="_get_user_badge_level")
+    silver_badge = fields.Integer('Silver badges count', compute="_get_user_badge_level")
+    bronze_badge = fields.Integer('Bronze badges count', compute="_get_user_badge_level")
+
+    @api.multi
+    def add_karma(self, karma):
+        for user in self:
+            user.karma += karma
+        return True
+
+    @api.multi
+    @api.depends('badge_ids')
+    def _get_user_badge_level(self):
+        """ Return total badge per level of users
+        TDE CLEANME: shouldn't check type is forum ? """
+        for user in self:
+            user.gold_badge = 0
+            user.silver_badge = 0
+            user.bronze_badge = 0
+
+        self.env.cr.execute("""
+            SELECT bu.user_id, b.level, count(1)
+            FROM gamification_badge_user bu, gamification_badge b
+            WHERE bu.user_id IN %s
+              AND bu.badge_id = b.id
+              AND b.level IS NOT NULL
+            GROUP BY bu.user_id, b.level
+            ORDER BY bu.user_id;
+        """, [tuple(self.ids)])
+
+        for (user_id, level, count) in self.env.cr.fetchall():
+            # levels are gold, silver, bronze but fields have _badge postfix
+            self.browse(user_id)['{}_badge'.format(level)] = count
