+++ addons/gamification/views/badge.xml
@@ -73,6 +73,7 @@
                     </group>
                     <group string="Rewards for challenges">
                         <field name="challenge_ids" mode="kanban" widget="many2many" nolabel="1" context="{'default_reward_id': active_id}" />
+                        <field name="level"/>
                     </group>
                     <group string="Statistics">
                         <group>
